requirements.txt
streamlit
pandas
plotly
seaborn
matplotlib
pillow
numpy